parallelize perf test

+++ perf/same_take.js
@@ -4,11 +4,14 @@ const rimraf       = require('rimraf');
 const _            = require('lodash');
 const Stats        = require('fast-stats').Stats;
 const Table        = require('cli-table');
-const ProgressBar  = require('progress');
 const spawn        = require('child_process').spawn;
+const cluster      = require('cluster');
+const os = require('os');
 
-(function spawn_server() {
-  console.log('spawning server');
+const requests = 100000;
+const concurrency = 1000;
+
+function spawn_server() {
 
   try{
     const db_file = path.join(__dirname, 'db', 'perf.tests.db');
@@ -23,15 +26,60 @@ const spawn        = require('child_process').spawn;
     limitd_args.push('--profile');
   }
 
-  const server = spawn(`${path.normalize(__dirname + '/../bin/limitd')}`, limitd_args, { stdio: 'inherit' });
+  return spawn(`${path.normalize(__dirname + '/../bin/limitd')}`, limitd_args, { stdio: 'inherit' });
+}
 
-  process.once('exit', (code) => {
-    server.kill();
-    process.exit(code);
+function render_results(started_at, results) {
+  var took    = new Date() - started_at;
+  var errored = _.filter(results, 'err');
+
+  var times = _(results).filter(function (r) { return !r.err; }).map('took').value();
+  var stats = new Stats().push(times);
+
+  var table = new Table();
+
+  table.push(
+      { 'Requests':    requests * os.cpus().length     },
+      { 'Concurrency': concurrency * os.cpus().length  },
+      { 'Total time':  took + ' ms'                    },
+      { 'Errored':     errored.length                  },
+      { 'Mean':        stats.amean().toFixed(2)        },
+      { 'P50':         stats.percentile(50).toFixed(2) },
+      { 'P95':         stats.percentile(95).toFixed(2) },
+      { 'P97':         stats.percentile(97).toFixed(2) },
+      { 'Max':         _.max(times)                    },
+      { 'Min':         _.min(times)                    }
+  );
+
+  console.log(table.toString());
+}
+
+if (cluster.isMaster) {
+  process.title = 'limitd perfomance master';
+
+  const results = [];
+  const cores = os.cpus().length;
+  const started_at = new Date();
+
+  const server = spawn_server();
+
+  const workers = _.range(cores).map(() => cluster.fork());
+
+
+  workers.forEach((worker) => {
+    worker.once('message', (message) => {
+      results.push(message.results);
+      if (results.length === cores) {
+        render_results(started_at, _.flatten(results));
+        server.kill();
+      }
+      worker.kill();
+    });
   });
-})();
 
-process.title = 'limitd perfomance client';
+  return;
+}
+
 
 const LimitdClient = require('limitd-client');
 
@@ -46,19 +94,12 @@ function waitAll(){
 }
 
 function run_tests () {
-  const started_at = new Date();
-
-  const requests = 200000;
-  const concurrency = 5000;
-
-  const progress = new ProgressBar(':bar', { total: requests , width: 50 });
-
   async.mapLimit(_.range(requests), concurrency, function (i, done) {
     const date = new Date();
     const client = clients.shift();
     clients.push(client);
-    return client.take('ip', '10.0.0.1', function (err, result) {
-      progress.tick();
+    // return client.take('ip', '10.0.0.1', function (err, result) {
+    return client.take('ip', `${process.pid}-${i}-pipi`, function (err, result) {
       if (err) {
         console.dir(err);
         return process.exit(1);
@@ -76,31 +117,10 @@ function run_tests () {
       return process.exit(1);
     }
 
-    var took    = new Date() - started_at;
-    var errored = _.filter(results, 'err');
-
-    var times = _(results).filter(function (r) { return !r.err; }).map('took').value();
-    var stats = new Stats().push(times);
-
-    var table = new Table();
-
-
-    table.push(
-        { 'Requests':   requests                        },
-        { 'Total time': took + ' ms'                    },
-        { 'Errored':    errored.length                  },
-        { 'Mean':       stats.amean().toFixed(2)        },
-        { 'P50':        stats.percentile(50).toFixed(2) },
-        { 'P95':        stats.percentile(95).toFixed(2) },
-        { 'P97':        stats.percentile(97).toFixed(2) },
-        { 'Max':        _.max(times)                    },
-        { 'Min':        _.min(times)                    }
-    );
-
-    console.log(table.toString());
-
-    process.exit(0);
-
+    process.send({
+      type: 'finish',
+      results
+    });
   });
 }
